+++ framework/console/controllers/MessageController.php
@@ -8,8 +8,10 @@
 namespace yii\console\controllers;
 
 use Yii;
+use yii\base\ErrorException;
 use yii\console\Controller;
 use yii\console\Exception;
+use yii\helpers\Console;
 use yii\helpers\FileHelper;
 use yii\helpers\VarDumper;
 use yii\i18n\GettextPoFile;
@@ -203,10 +205,10 @@ class MessageController extends Controller
 
                 $db->createCommand()
                    ->insert($sourceMessageTable, ['category' => $category, 'message' => $m])->execute();
-                $lastId = $db->getLastInsertID();
+                $lastID = $db->getLastInsertID();
                 foreach ($languages as $language) {
                     $db->createCommand()
-                       ->insert($messageTable, ['id' => $lastId, 'language' => $language])->execute();
+                       ->insert($messageTable, ['id' => $lastID, 'language' => $language])->execute();
                 }
             }
         }
@@ -222,17 +224,12 @@ class MessageController extends Controller
                    ->delete($sourceMessageTable, ['in', 'id', $obsolete])->execute();
                 echo "deleted.\n";
             } else {
-                $last_id = $db->getLastInsertID();
                 $db->createCommand()
                    ->update(
                        $sourceMessageTable,
                        ['message' => new \yii\db\Expression("CONCAT('@@',message,'@@')")],
                        ['in', 'id', $obsolete]
                    )->execute();
-                foreach ($languages as $language) {
-                    $db->createCommand()
-                       ->insert($messageTable, ['id' => $last_id, 'language' => $language])->execute();
-                }
                 echo "updated.\n";
             }
         }
@@ -263,7 +260,17 @@ class MessageController extends Controller
             for ($i = 0; $i < $n; ++$i) {
                 $category = substr($matches[$i][1], 1, -1);
                 $message = $matches[$i][2];
-                $messages[$category][] = eval("return {$message};"); // use eval to eliminate quote escape
+                try {
+                    $messages[$category][] = eval("return {$message};"); // use eval to eliminate quote escape
+                } catch (ErrorException $e) {
+                    $category = Console::ansiFormat($category, [Console::FG_CYAN]);
+                    $message = Console::ansiFormat($message, [Console::FG_CYAN]);
+                    $fileName = Console::ansiFormat($fileName, [Console::FG_CYAN]);
+                    $error = Console::ansiFormat($e->getMessage(), [Console::FG_RED]);
+
+                    $this->stdout("Failed parsing $fileName, $message in $category category:\n" . $error . "\n");
+                    Yii::$app->end(self::EXIT_CODE_ERROR);
+                }
             }
         }