+++ tests/unit/extensions/mongodb/ActiveRecordTest.php
@@ -263,4 +263,16 @@ class ActiveRecordTest extends MongoDbTestCase
         $this->assertNotEmpty($rowRefreshed);
         $this->assertEquals(7, $rowRefreshed->status);
     }
+
+    public function testModify()
+    {
+        $searchName = 'name7';
+        $newName = 'new name';
+
+        $customer = Customer::find()
+            ->where(['name' => $searchName])
+            ->modify(['$set' => ['name' => $newName]], ['new' => true]);
+        $this->assertTrue($customer instanceof Customer);
+        $this->assertEquals($newName, $customer->name);
+    }
 }