+++ extensions/mongodb/CHANGELOG.md
@@ -5,6 +5,7 @@ Yii Framework 2 mongodb extension Change Log
 -----------------------
 
 - Enh #3855: Added debug toolbar panel for MongoDB (klimov-paul)
+- Enh #5592: Added support for 'findAndModify' operation at `yii\mongodb\Query` and `yii\mongodb\ActiveQuery` (klimov-paul)
 
 
 2.0.0 October 12, 2014