+++ extensions/mongodb/CHANGELOG.md
@@ -1,7 +1,7 @@
 Yii Framework 2 mongodb extension Change Log
 ============================================
 
-2.0.1 under development
+2.0.1 December 07, 2014
 -----------------------
 
 - Bug #6026: Fixed `yii\mongodb\ActiveRecord` saves `null` as `_id`, if attributes are empty (klimov-paul)